renamed linter -> Linter
